+++ src/theory/strings/theory_strings.cpp
@@ -78,6 +78,7 @@ TheoryStrings::TheoryStrings(Env& env, OutputChannel& out, Valuation valuation)
                 d_csolver,
                 d_extTheory,
                 d_statistics),
+      d_psolver(env, d_state, d_im, d_termReg, d_bsolver),
       d_asolver(env,
                 d_state,
                 d_im,
@@ -1106,113 +1107,6 @@ void TheoryStrings::notifySharedTerm(TNode n)
   }
 }
 
-void TheoryStrings::checkCodes()
-{
-  // ensure that lemmas regarding str.code been added for each constant string
-  // of length one
-  if (d_termReg.hasStringCode())
-  {
-    NodeManager* nm = NodeManager::currentNM();
-    // str.code applied to the code term for each equivalence class that has a
-    // code term but is not a constant
-    std::vector<Node> nconst_codes;
-    // str.code applied to the proxy variables for each equivalence classes that
-    // are constants of size one
-    std::vector<Node> const_codes;
-    const std::vector<Node>& seqc = d_bsolver.getStringLikeEqc();
-    for (const Node& eqc : seqc)
-    {
-      if (!eqc.getType().isString())
-      {
-        continue;
-      }
-      if (eqc.isConst())
-      {
-        Node c = eqc;
-        Trace("strings-code-debug") << "Get proxy variable for " << c
-                                    << std::endl;
-        Node cc = nm->mkNode(kind::STRING_TO_CODE, c);
-        cc = rewrite(cc);
-        Assert(cc.isConst());
-        Node cp = d_termReg.ensureProxyVariableFor(c);
-        Node vc = nm->mkNode(STRING_TO_CODE, cp);
-        if (!d_state.areEqual(cc, vc))
-        {
-          std::vector<Node> emptyVec;
-          d_im.sendInference(
-              emptyVec, cc.eqNode(vc), InferenceId::STRINGS_CODE_PROXY);
-        }
-        // only relevant for injectivity if length is 1 (e.g. it has a valid
-        // code point)
-        if (Word::getLength(c) == 1)
-        {
-          const_codes.push_back(vc);
-        }
-      }
-      else
-      {
-        EqcInfo* ei = d_state.getOrMakeEqcInfo(eqc, false);
-        if (ei && !ei->d_codeTerm.get().isNull())
-        {
-          Node vc = nm->mkNode(kind::STRING_TO_CODE, ei->d_codeTerm.get());
-          // only relevant for injectivity if not already equal to negative one
-          if (!d_state.areEqual(vc, d_neg_one))
-          {
-            nconst_codes.push_back(vc);
-          }
-        }
-      }
-    }
-    if (d_im.hasProcessed())
-    {
-      return;
-    }
-    // now, ensure that str.code is injective
-    std::vector<Node> cmps;
-    cmps.insert(cmps.end(), const_codes.rbegin(), const_codes.rend());
-    cmps.insert(cmps.end(), nconst_codes.rbegin(), nconst_codes.rend());
-    for (unsigned i = 0, num_ncc = nconst_codes.size(); i < num_ncc; i++)
-    {
-      Node c1 = nconst_codes[i];
-      cmps.pop_back();
-      for (const Node& c2 : cmps)
-      {
-        Trace("strings-code-debug")
-            << "Compare codes : " << c1 << " " << c2 << std::endl;
-        if (!d_state.areDisequal(c1, c2))
-        {
-          Node eq_no = c1.eqNode(d_neg_one);
-          Node deq = c1.eqNode(c2).negate();
-          Node eqn = c1[0].eqNode(c2[0]);
-          // str.code(x)==-1 V str.code(x)!=str.code(y) V x==y
-          Node inj_lem = nm->mkNode(kind::OR, eq_no, deq, eqn);
-          deq = rewrite(deq);
-          d_im.addPendingPhaseRequirement(deq, false);
-          std::vector<Node> emptyVec;
-          d_im.sendInference(emptyVec, inj_lem, InferenceId::STRINGS_CODE_INJ);
-        }
-      }
-    }
-  }
-}
-
-void TheoryStrings::checkRegisterTermsNormalForms()
-{
-  const std::vector<Node>& seqc = d_bsolver.getStringLikeEqc();
-  for (const Node& eqc : seqc)
-  {
-    NormalForm& nfi = d_csolver.getNormalForm(eqc);
-    // check if there is a length term for this equivalence class
-    EqcInfo* ei = d_state.getOrMakeEqcInfo(eqc, false);
-    Node lt = ei ? ei->d_lengthTerm : Node::null();
-    if (lt.isNull())
-    {
-      Node c = d_termReg.mkNConcat(nfi.d_nf, eqc.getType());
-      d_termReg.registerTerm(c);
-    }
-  }
-}
-
 TrustNode TheoryStrings::ppRewrite(TNode atom, std::vector<SkolemLemma>& lems)
 {
   Trace("strings-ppr") << "TheoryStrings::ppRewrite " << atom << std::endl;
@@ -1306,12 +1200,12 @@ void TheoryStrings::runInferStep(InferStep s, Theory::Effort e, int effort)
     case CHECK_FLAT_FORMS: d_csolver.checkFlatForms(); break;
     case CHECK_NORMAL_FORMS_EQ: d_csolver.checkNormalFormsEq(); break;
     case CHECK_NORMAL_FORMS_DEQ: d_csolver.checkNormalFormsDeq(); break;
-    case CHECK_CODES: checkCodes(); break;
+    case CHECK_CODES: d_psolver.checkCodes(); break;
     case CHECK_LENGTH_EQC: d_csolver.checkLengthsEqc(); break;
     case CHECK_SEQUENCES_ARRAY_CONCAT: d_asolver.checkArrayConcat(); break;
     case CHECK_SEQUENCES_ARRAY: d_asolver.checkArray(); break;
     case CHECK_SEQUENCES_ARRAY_EAGER: d_asolver.checkArrayEager(); break;
-    case CHECK_REGISTER_TERMS_NF: checkRegisterTermsNormalForms(); break;
+    case CHECK_REGISTER_TERMS_NF: d_csolver.checkRegisterTermsNormalForms(); break;
     case CHECK_EXTF_REDUCTION: d_esolver.checkExtfReductions(effort); break;
     case CHECK_MEMBERSHIP: d_rsolver.checkMemberships(effort); break;
     case CHECK_CARDINALITY: d_bsolver.checkCardinality(); break;
