+++ src/theory/strings/theory_strings.h
@@ -28,6 +28,7 @@
 #include "theory/ext_theory.h"
 #include "theory/strings/array_solver.h"
 #include "theory/strings/base_solver.h"
+#include "theory/strings/code_point_solver.h"
 #include "theory/strings/core_solver.h"
 #include "theory/strings/eager_solver.h"
 #include "theory/strings/extf_solver.h"
@@ -197,30 +198,6 @@ class TheoryStrings : public Theory {
    * of atom, including calls to registerTerm.
    */
   void assertPendingFact(Node atom, bool polarity, Node exp);
-  //-----------------------inference steps
-  /** check register terms pre-normal forms
-   *
-   * This calls registerTerm(n,2) on all non-congruent strings in the
-   * equality engine of this class.
-   */
-  void checkRegisterTermsPreNormalForm();
-  /** check codes
-   *
-   * This inference schema ensures that constraints between str.code terms
-   * are satisfied by models that correspond to extensions of the current
-   * assignment. In particular, this method ensures that str.code can be
-   * given an interpretation that is injective for string arguments with length
-   * one. It may add lemmas of the form:
-   *   str.code(x) == -1 V str.code(x) != str.code(y) V x == y
-   */
-  void checkCodes();
-  /** check register terms for normal forms
-   *
-   * This calls registerTerm(str.++(t1, ..., tn ), 3) on the normal forms
-   * (t1, ..., tn) of all string equivalence classes { s1, ..., sm } such that
-   * there does not exist a term of the form str.len(si) in the current context.
-   */
-  void checkRegisterTermsNormalForms();
   /**
    * Turn a sequence constant into a skeleton specifying how to construct
    * its value.
@@ -298,6 +275,8 @@ class TheoryStrings : public Theory {
    * involving extended string functions.
    */
   ExtfSolver d_esolver;
+  /** Code point solver */
+  CodePointSolver d_psolver;
   /**
    * The array solver, which implements specialized approaches for
    * seq.nth/seq.update.
